+++ src/utils/executer.ts
@@ -16,3 +16,33 @@ export async function execute(
     });
   });
 }
+
+export const initNpmRepository = async (projectPath: string) => {
+  // Initialize the npm repository
+  try {
+    console.log('Initializing npm repository...');
+    await execute('npm init -y', projectPath, {
+      log: false,
+      cwd: projectPath,
+    });
+    console.log('npm repository successfully initialized.');
+
+    // await installHardhat(projectPath);
+  } catch (e: any) {
+    console.error(e.message);
+  }
+};
+
+export const installHardhat = async (projectPath: string) => {
+  try {
+    console.log('Installing Hardhat...');
+    await execute(`npm install --save-dev hardhat`, projectPath, {
+      log: false,
+      cwd: projectPath,
+    });
+  } catch (e) {
+    console.error(e);
+  }
+
+  console.log('Hardhat successfully installed.');
+};
